Implement requests-cache for RawClient
